refactor: remove unused ui component
