--- src/components/Button.tsx
@@ -1,218 +1,0 @@
-/* eslint-disable react/display-name */
-import clsx from 'clsx';
-import React from 'react';
-import { type IconType } from 'react-icons';
-
-import BasicLink, { type BasicLinkProps } from './BasicLink';
-
-type ButtonVariant =
-  | 'filled'
-  | 'outlined'
-  | 'text'
-  | 'elevated'
-  | 'tonal'
-  | 'danger';
-type ButtonSize = 'sm' | 'md' | 'lg';
-
-type ButtonProps = {
-  variant: ButtonVariant;
-  size?: ButtonSize;
-  icon?: IconType;
-  iconClassName?: string;
-} & React.ComponentPropsWithRef<'button'>;
-
-export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  (
-    {
-      children,
-      className,
-      disabled: buttonDisabled,
-      variant = 'filled',
-      size = 'md',
-      icon: Icon,
-      iconClassName,
-      ...rest
-    },
-    ref,
-  ) => {
-    const disabled = buttonDisabled ?? false;
-    return (
-      <button
-        ref={ref}
-        type='button'
-        disabled={disabled}
-        className={clsx(
-          'inline-flex items-center rounded-3xl font-medium hover:shadow-md disabled:cursor-not-allowed disabled:text-neutral-500 disabled:shadow-none',
-          'focus:outline-none focus-visible:ring focus-visible:ring-primary-500',
-          'transition-colors duration-75',
-          className,
-          [
-            size === 'sm' && 'px-2 py-1 text-sm',
-            size === 'md' && 'px-3 py-2 text-sm',
-            size === 'lg' && 'px-4 py-3 text-base',
-          ],
-          [
-            variant === 'filled' &&
-              'bg-primary-500 text-white hover:bg-primary-400 disabled:bg-neutral-900 ',
-            variant === 'danger' &&
-              'bg-error-500 text-white hover:bg-error-400 disabled:bg-neutral-900 ',
-            variant === 'outlined' &&
-              'border border-neutral-600 bg-transparent text-primary-400 hover:border-primary-500 hover:bg-neutral-900 disabled:border-neutral-600 disabled:bg-transparent',
-            variant === 'text' &&
-              'text-primary-400 hover:bg-neutral-900 disabled:bg-transparent',
-            variant === 'elevated' &&
-              'bg-slate-100 text-primary-400 shadow-md shadow-slate-400 hover:bg-slate-200 disabled:bg-neutral-900 disabled:shadow-none',
-            variant === 'tonal' &&
-              'bg-secondary-900 hover:bg-secondary-800 disabled:bg-neutral-900',
-          ],
-        )}
-        {...rest}
-      >
-        {Icon && (
-          <span className='mr-2'>
-            <Icon
-              className={clsx(
-                [
-                  size === 'sm' && 'text-sm',
-                  size === 'md' && 'text-sm',
-                  size === 'lg' && 'text-base',
-                ],
-                iconClassName,
-              )}
-            />
-          </span>
-        )}
-        {children}
-      </button>
-    );
-  },
-);
-
-type LinkButtonProps = {
-  variant: ButtonVariant;
-  size?: ButtonSize;
-  icon?: IconType;
-  iconClassName?: string;
-} & BasicLinkProps;
-
-export const LinkButton = React.forwardRef<HTMLAnchorElement, LinkButtonProps>(
-  (
-    {
-      children,
-      className,
-      variant = 'filled',
-      size = 'md',
-      icon: Icon,
-      iconClassName,
-      ...rest
-    },
-    ref,
-  ) => {
-    return (
-      <BasicLink
-        ref={ref}
-        {...rest}
-        className={clsx(
-          'inline-flex items-center rounded-3xl font-medium hover:shadow-md disabled:cursor-not-allowed disabled:text-neutral-500 disabled:shadow-none',
-          'focus:outline-none focus-visible:ring focus-visible:ring-primary-500',
-          'transition-colors duration-75',
-          [
-            size === 'sm' && 'px-2 py-1 text-sm',
-            size === 'md' && 'px-3 py-2 text-sm',
-            size === 'lg' && 'px-4 py-3 text-base',
-          ],
-          [
-            variant === 'filled' &&
-              'bg-primary-500 text-white hover:bg-primary-400 disabled:bg-neutral-900 ',
-            variant === 'danger' &&
-              'bg-error-500 text-white hover:bg-error-400 disabled:bg-neutral-900 ',
-            variant === 'outlined' &&
-              'border border-neutral-600 bg-transparent text-primary-400 hover:border-primary-500 hover:bg-neutral-900 disabled:border-neutral-600 disabled:bg-transparent',
-            variant === 'text' &&
-              'text-primary-400 hover:bg-neutral-900 disabled:bg-transparent',
-            variant === 'elevated' &&
-              'bg-slate-100 text-primary-400 shadow-md shadow-slate-400 hover:bg-slate-200 disabled:bg-neutral-900 disabled:shadow-none',
-            variant === 'tonal' &&
-              'bg-secondary-900 hover:bg-secondary-800 disabled:bg-neutral-900',
-          ],
-          className,
-        )}
-      >
-        {Icon && (
-          <span className='mr-2'>
-            <Icon
-              className={clsx(
-                [
-                  size === 'sm' && 'text-sm',
-                  size === 'md' && 'text-sm',
-                  size === 'lg' && 'text-base',
-                ],
-                iconClassName,
-              )}
-            />
-          </span>
-        )}
-        {children}
-      </BasicLink>
-    );
-  },
-);
-
-type IconButtonProps = {
-  variant: ButtonVariant;
-  size?: ButtonSize;
-  icon: IconType;
-  iconClassname?: string;
-} & React.ComponentPropsWithRef<'button'>;
-
-export const IconButton = React.forwardRef<HTMLButtonElement, IconButtonProps>(
-  (
-    {
-      className,
-      disabled: buttonDisabled,
-      variant = 'filled',
-      size = 'md',
-      icon: Icon,
-      iconClassname,
-      ...rest
-    },
-    ref,
-  ) => {
-    const disabled = buttonDisabled ?? false;
-    return (
-      <button
-        ref={ref}
-        type='button'
-        disabled={disabled}
-        className={clsx(
-          'inline-flex items-center rounded-full font-medium hover:shadow-md disabled:cursor-not-allowed disabled:text-neutral-500 disabled:shadow-none',
-          'focus:outline-none focus-visible:ring focus-visible:ring-primary-500',
-          'transition-colors duration-75',
-          [
-            size === 'sm' && 'p-1 text-xs',
-            size === 'md' && 'p-2.5 text-base',
-            size === 'lg' && 'p-3 text-lg',
-          ],
-          [
-            variant === 'filled' &&
-              'bg-primary-500 text-white hover:bg-primary-400 disabled:bg-neutral-900 ',
-            variant === 'danger' &&
-              'bg-error-500 text-white hover:bg-error-400 disabled:bg-neutral-900 ',
-            variant === 'outlined' &&
-              'border border-neutral-600 bg-transparent text-primary-400 hover:border-primary-500 hover:bg-neutral-900 disabled:border-neutral-600 disabled:bg-transparent',
-            variant === 'text' &&
-              'text-primary-400 hover:bg-neutral-900 disabled:bg-transparent',
-            variant === 'elevated' &&
-              'bg-slate-100 text-primary-400 shadow-md shadow-slate-400 hover:bg-slate-200 disabled:bg-neutral-900 disabled:shadow-none',
-            variant === 'tonal' &&
-              'bg-secondary-900 hover:bg-secondary-800 disabled:bg-neutral-900',
-          ],
-          className,
-        )}
-        {...rest}
-      >
-        {Icon && <Icon className={iconClassname} />}
-      </button>
-    );
-  },
-);
